YDM-8 / watched movies list implementation

+++ src/main/java/pl/szczesniak/dominik/whattowatch/movies/domain/MoviesToWatchService.java
@@ -4,6 +4,7 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import pl.szczesniak.dominik.whattowatch.movies.domain.model.MovieId;
 import pl.szczesniak.dominik.whattowatch.movies.domain.model.MovieTitle;
+import pl.szczesniak.dominik.whattowatch.movies.domain.model.exceptions.MovieDoesNotExistException;
 import pl.szczesniak.dominik.whattowatch.movies.domain.model.exceptions.UserDoesNotExistException;
 import pl.szczesniak.dominik.whattowatch.users.domain.model.UserId;
 
@@ -33,14 +34,17 @@ public class MoviesToWatchService {
 
 	public void moveMovieToWatchedList(MovieId movieId, UserId userId) {
 		userCheck(userId);
-		final WatchedMovie watchedMovie = new WatchedMovie(movieId, repository.getMovieTitle(movieId), userId);
+		if (repository.findBy(movieId, userId).isEmpty()) {
+			throw new MovieDoesNotExistException("No movie matched movieId: " + movieId + " and userId: " + userId + ". Action aborted.");
+		}
+		final WatchedMovie watchedMovie = new WatchedMovie(movieId, repository.getMovie(movieId, userId).getTitle(), userId);
 		watchedRepository.addMovieToWatchedList(watchedMovie);
 		repository.removeMovie(movieId, userId);
 	}
 
 	private void userCheck(final UserId userId) {
 		if (!userProvider.exists(userId)) {
-			throw new UserDoesNotExistException("User doesn't exist. Action aborted");
+			throw new UserDoesNotExistException("User with userId: " + userId + " doesn't exist. Action aborted");
 		}
 	}
